feat: command permissions
Every command action can have its own permission or set of permissions required to execute the command.

You can use a ! symbol in front of the permission to invert the condition, making the permission strictly not required to execute the command. If they have that permission and its not set to false then they will not be allowed to execute the command.