fix: always-give itemflag
The proper function of the always-give itemflag is to ALWAYS give the item upon performing a trigger or getting the item via commands regardless of the player already having the item or not.
This wasn't being respected and was not giving the item in the designated slot FIRST before adding it to the next available if the designated slot was occupied.